Release 1.25.1203.1

+++ Modules/Microsoft365DSC/DSCResources/MSFT_TeamsOrgWideAppSettings/MSFT_TeamsOrgWideAppSettings.psm1
@@ -20,12 +20,24 @@ function Get-TargetResource
         $Credential,
 
         [Parameter()]
-        [System.String[]]
-        $AccessTokens,
+        [System.String]
+        $ApplicationId,
+
+        [Parameter()]
+        [System.String]
+        $TenantId,
+
+        [Parameter()]
+        [System.String]
+        $CertificateThumbprint,
 
         [Parameter()]
         [Switch]
-        $ManagedIdentity
+        $ManagedIdentity,
+
+        [Parameter()]
+        [System.String[]]
+        $AccessTokens
     )
 
     Write-Verbose -Message 'Checking the Teams Org Wide App Settings'
@@ -55,9 +67,11 @@ function Get-TargetResource
         return @{
             IsSingleInstance                   = 'Yes'
             IsSideloadedAppsInteractionEnabled = $settings.IsSideloadedAppsInteractionEnabled
-            Credential                         = $Credential
-            AccessTokens                       = $AccessTokens
+            ApplicationId                      = $ApplicationId
+            TenantId                           = $TenantId
+            CertificateThumbprint              = $CertificateThumbprint
             ManagedIdentity                    = $ManagedIdentity.IsPresent
+            AccessTokens                       = $AccessTokens
         }
     }
     catch
@@ -100,12 +114,24 @@ function Set-TargetResource
         $Credential,
 
         [Parameter()]
-        [System.String[]]
-        $AccessTokens,
+        [System.String]
+        $ApplicationId,
+
+        [Parameter()]
+        [System.String]
+        $TenantId,
+
+        [Parameter()]
+        [System.String]
+        $CertificateThumbprint,
 
         [Parameter()]
         [Switch]
-        $ManagedIdentity
+        $ManagedIdentity,
+
+        [Parameter()]
+        [System.String[]]
+        $AccessTokens
     )
 
     Write-Verbose -Message 'Setting the Teams Org Wide App Settings'
@@ -125,8 +151,15 @@ function Set-TargetResource
     $null = New-M365DSCConnection -Workload 'MicrosoftTeams' `
         -InboundParameters $PSBoundParameters
 
-    $SetParameters = Remove-M365DSCAuthenticationParameter -BoundParameters $PSBoundParameters
-    Set-CsTeamsSettingsCustomApp @SetParameters
+    if ($PSBoundParameters.ContainsKey('ApplicationId'))
+    {
+        Write-Warning -Message "This resources doesn't support 'write' operations using Service Principal Authentication."
+    }
+    else
+    {
+        $SetParameters = Remove-M365DSCAuthenticationParameter -BoundParameters $PSBoundParameters
+        Set-CsTeamsSettingsCustomApp @SetParameters
+    }
 }
 
 function Test-TargetResource
@@ -149,12 +182,24 @@ function Test-TargetResource
         $Credential,
 
         [Parameter()]
-        [System.String[]]
-        $AccessTokens,
+        [System.String]
+        $ApplicationId,
+
+        [Parameter()]
+        [System.String]
+        $TenantId,
+
+        [Parameter()]
+        [System.String]
+        $CertificateThumbprint,
 
         [Parameter()]
         [Switch]
-        $ManagedIdentity
+        $ManagedIdentity,
+
+        [Parameter()]
+        [System.String[]]
+        $AccessTokens
     )
 
     #region Telemetry
@@ -182,6 +227,22 @@ function Export-TargetResource
         $Credential,
 
         [Parameter()]
+        [System.String]
+        $ApplicationId,
+
+        [Parameter()]
+        [System.String]
+        $TenantId,
+
+        [Parameter()]
+        [System.String]
+        $CertificateThumbprint,
+
+        [Parameter()]
+        [Switch]
+        $ManagedIdentity,
+
+        [Parameter()]
         [System.String[]]
         $AccessTokens
     )
@@ -205,27 +266,30 @@ function Export-TargetResource
     {
         $dscContent = ''
         $params = @{
-            IsSingleInstance = 'Yes'
-            Credential       = $Credential
-            AccessTokens     = $AccessTokens
+            IsSingleInstance      = 'Yes'
+            Credential            = $Credential
+            ApplicationId         = $ApplicationId
+            TenantId              = $TenantId
+            CertificateThumbprint = $CertificateThumbprint
+            ManagedIdentity       = $ManagedIdentity.IsPresent
+            AccessTokens          = $AccessTokens
         }
-        $Results = Get-TargetResource @Params
-        if ($Results.Ensure -eq 'Present')
-        {
-            if ($null -ne $Global:M365DSCExportResourceInstancesCount)
-            {
-                $Global:M365DSCExportResourceInstancesCount++
-            }
 
-            $currentDSCBlock = Get-M365DSCExportContentForResource -ResourceName $ResourceName `
-                -ConnectionMode $ConnectionMode `
-                -ModulePath $PSScriptRoot `
-                -Results $Results `
-                -Credential $Credential
-            $dscContent += $currentDSCBlock
-            Save-M365DSCPartialExport -Content $currentDSCBlock `
-                -FileName $Global:PartialExportFileName
+        $Results = Get-TargetResource @params
+        if ($null -ne $Global:M365DSCExportResourceInstancesCount)
+        {
+            $Global:M365DSCExportResourceInstancesCount++
         }
+
+        $currentDSCBlock = Get-M365DSCExportContentForResource -ResourceName $ResourceName `
+            -ConnectionMode $ConnectionMode `
+            -ModulePath $PSScriptRoot `
+            -Results $Results `
+            -Credential $Credential
+
+        $dscContent += $currentDSCBlock
+        Save-M365DSCPartialExport -Content $currentDSCBlock `
+            -FileName $Global:PartialExportFileName
 
         Write-M365DSCHost -Message $Global:M365DSCEmojiGreenCheckMark -CommitWrite
         return $dscContent
